requirements.txt
Django==5.0
djangorestframework==3.14.0
qrcode==7.4.2
pillow==10.1.0
django-bootstrap5==23.3
